Challenges/ISPctf/darkmagic/requirements.txt
urllib3
flask
selenium
webdriver_manager
Pillow